Refactored Storage Class By Splitting Into 3 Different Classes
