+++ src/main/java/javatro/core/PlanetCard.java
@@ -1,6 +1,7 @@
 // @@author Markneoneo
 package javatro.core;
 
+import javatro.storage.DataParser;
 import javatro.storage.Storage;
 
 import java.util.EnumMap;
@@ -36,19 +37,19 @@ public class PlanetCard {
             Storage storage = Storage.getStorageInstance();
             if (storage.getRunChosen() < storage.getNumberOfRuns()) {
                 // Saved Run, load data from csv
-                LEVELS.put(PokerHand.HandType.HIGH_CARD, Storage.HIGH_CARD_INDEX);
-                LEVELS.put(PokerHand.HandType.PAIR, Storage.PAIR_INDEX);
-                LEVELS.put(PokerHand.HandType.TWO_PAIR, Storage.TWO_PAIR_INDEX);
-                LEVELS.put(PokerHand.HandType.THREE_OF_A_KIND, Storage.THREE_OF_A_KIND_INDEX);
-                LEVELS.put(PokerHand.HandType.STRAIGHT, Storage.STRAIGHT_INDEX);
-                LEVELS.put(PokerHand.HandType.FLUSH, Storage.FLUSH_INDEX);
-                LEVELS.put(PokerHand.HandType.FULL_HOUSE, Storage.FULL_HOUSE_INDEX);
-                LEVELS.put(PokerHand.HandType.FOUR_OF_A_KIND, Storage.FOUR_OF_A_KIND_INDEX);
-                LEVELS.put(PokerHand.HandType.STRAIGHT_FLUSH, Storage.STRAIGHT_FLUSH_INDEX);
-                LEVELS.put(PokerHand.HandType.ROYAL_FLUSH, Storage.ROYAL_FLUSH_INDEX);
-                LEVELS.put(PokerHand.HandType.FIVE_OF_A_KIND, Storage.FIVE_OF_A_KIND_INDEX);
-                LEVELS.put(PokerHand.HandType.FLUSH_HOUSE, Storage.FLUSH_HOUSE_INDEX);
-                LEVELS.put(PokerHand.HandType.FLUSH_FIVE, Storage.FLUSH_FIVE_INDEX);
+                LEVELS.put(PokerHand.HandType.HIGH_CARD, DataParser.HIGH_CARD_INDEX);
+                LEVELS.put(PokerHand.HandType.PAIR, DataParser.PAIR_INDEX);
+                LEVELS.put(PokerHand.HandType.TWO_PAIR, DataParser.TWO_PAIR_INDEX);
+                LEVELS.put(PokerHand.HandType.THREE_OF_A_KIND, DataParser.THREE_OF_A_KIND_INDEX);
+                LEVELS.put(PokerHand.HandType.STRAIGHT, DataParser.STRAIGHT_INDEX);
+                LEVELS.put(PokerHand.HandType.FLUSH, DataParser.FLUSH_INDEX);
+                LEVELS.put(PokerHand.HandType.FULL_HOUSE, DataParser.FULL_HOUSE_INDEX);
+                LEVELS.put(PokerHand.HandType.FOUR_OF_A_KIND, DataParser.FOUR_OF_A_KIND_INDEX);
+                LEVELS.put(PokerHand.HandType.STRAIGHT_FLUSH, DataParser.STRAIGHT_FLUSH_INDEX);
+                LEVELS.put(PokerHand.HandType.ROYAL_FLUSH, DataParser.ROYAL_FLUSH_INDEX);
+                LEVELS.put(PokerHand.HandType.FIVE_OF_A_KIND, DataParser.FIVE_OF_A_KIND_INDEX);
+                LEVELS.put(PokerHand.HandType.FLUSH_HOUSE, DataParser.FLUSH_HOUSE_INDEX);
+                LEVELS.put(PokerHand.HandType.FLUSH_FIVE, DataParser.FLUSH_FIVE_INDEX);
             } else {
                 LEVELS.put(handType, 1);
             }
@@ -208,55 +209,55 @@ public class PlanetCard {
         // Update Planet Cards
         storage.setValue(
                 storage.getRunChosen() - 1,
-                Storage.HIGH_CARD_INDEX,
+                DataParser.HIGH_CARD_INDEX,
                 String.valueOf(PlanetCard.getLevel(PokerHand.HandType.HIGH_CARD)));
         storage.setValue(
                 storage.getRunChosen() - 1,
-                Storage.PAIR_INDEX,
+                DataParser.PAIR_INDEX,
                 String.valueOf(PlanetCard.getLevel(PokerHand.HandType.PAIR)));
         storage.setValue(
                 storage.getRunChosen() - 1,
-                Storage.TWO_PAIR_INDEX,
+                DataParser.TWO_PAIR_INDEX,
                 String.valueOf(PlanetCard.getLevel(PokerHand.HandType.TWO_PAIR)));
         storage.setValue(
                 storage.getRunChosen() - 1,
-                Storage.THREE_OF_A_KIND_INDEX,
+                DataParser.THREE_OF_A_KIND_INDEX,
                 String.valueOf(PlanetCard.getLevel(PokerHand.HandType.THREE_OF_A_KIND)));
         storage.setValue(
                 storage.getRunChosen() - 1,
-                Storage.STRAIGHT_INDEX,
+                DataParser.STRAIGHT_INDEX,
                 String.valueOf(PlanetCard.getLevel(PokerHand.HandType.STRAIGHT)));
         storage.setValue(
                 storage.getRunChosen() - 1,
-                Storage.FLUSH_INDEX,
+                DataParser.FLUSH_INDEX,
                 String.valueOf(PlanetCard.getLevel(PokerHand.HandType.FLUSH)));
         storage.setValue(
                 storage.getRunChosen() - 1,
-                Storage.FULL_HOUSE_INDEX,
+                DataParser.FULL_HOUSE_INDEX,
                 String.valueOf(PlanetCard.getLevel(PokerHand.HandType.FULL_HOUSE)));
         storage.setValue(
                 storage.getRunChosen() - 1,
-                Storage.FOUR_OF_A_KIND_INDEX,
+                DataParser.FOUR_OF_A_KIND_INDEX,
                 String.valueOf(PlanetCard.getLevel(PokerHand.HandType.FOUR_OF_A_KIND)));
         storage.setValue(
                 storage.getRunChosen() - 1,
-                Storage.STRAIGHT_FLUSH_INDEX,
+                DataParser.STRAIGHT_FLUSH_INDEX,
                 String.valueOf(PlanetCard.getLevel(PokerHand.HandType.STRAIGHT_FLUSH)));
         storage.setValue(
                 storage.getRunChosen() - 1,
-                Storage.ROYAL_FLUSH_INDEX,
+                DataParser.ROYAL_FLUSH_INDEX,
                 String.valueOf(PlanetCard.getLevel(PokerHand.HandType.ROYAL_FLUSH)));
         storage.setValue(
                 storage.getRunChosen() - 1,
-                Storage.FIVE_OF_A_KIND_INDEX,
+                DataParser.FIVE_OF_A_KIND_INDEX,
                 String.valueOf(PlanetCard.getLevel(PokerHand.HandType.FIVE_OF_A_KIND)));
         storage.setValue(
                 storage.getRunChosen() - 1,
-                Storage.FLUSH_HOUSE_INDEX,
+                DataParser.FLUSH_HOUSE_INDEX,
                 String.valueOf(PlanetCard.getLevel(PokerHand.HandType.FLUSH_HOUSE)));
         storage.setValue(
                 storage.getRunChosen() - 1,
-                Storage.FLUSH_FIVE_INDEX,
+                DataParser.FLUSH_FIVE_INDEX,
                 String.valueOf(PlanetCard.getLevel(PokerHand.HandType.FLUSH_FIVE)));
 
         try {
